// src/app/api/models/iotHubDescription.ts
/***********************************************************
 * Copyright (c) Microsoft Corporation. All rights reserved.
 * Licensed under the MIT License
 **********************************************************/
export interface IotHubDescription {
    id: string;
    name: string;
    location: string;
}